*BUG FIX: When using Replication and Random Primary Keys, running the daily procedures report was causing and exception. *BUG FIX: When using Replication and Random Primary Keys, clicking the "Perio Chart" button was causing an exception sometimes. *BUG FIX: When using Replication and Random Primary Keys, an exception was always being caused when performing an image scan.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3142 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Main Modules/ContrDocs.cs
@@ -1615,7 +1615,7 @@ namespace OpenDental{
 				return;
 			}
 			DataRow obj=(DataRow)TreeDocuments.SelectedNode.Tag;
-			int docNum=Convert.ToInt32(obj["DocNum"].ToString());
+			long docNum=Convert.ToInt64(obj["DocNum"].ToString());
 			if(docNum!=0){//Only applied to document nodes.
 				if(!paintTools.Enabled){
 					EraseCurrentImages();
